refactor(SimuNEX): change static class to namespace. For StateSpace and ODESolvers.
