+++ UnityProject/Assets/Scripts/Core/Derived/Sensors/IdealMotorSensor.cs
@@ -1,4 +1,6 @@
 using System;
+using SimuNEX.Models;
+using SimuNEX.Solvers;
 using UnityEngine;
 
 namespace SimuNEX
@@ -62,16 +64,16 @@ namespace SimuNEX
         private StateSpace stateSpace;
 
         /// <summary>
-        /// The stepper method.
+        /// The solver method.
         /// </summary>
-        public StepperMethod stepper;
+        public SolverMethod solver;
 
         protected override void Initialize()
         {
             if (readPosition || readTorque)
             {
                 stateSpace = new();
-                stateSpace.Initialize(2, 1, new Matrix(2, 1), stepper: StateSpace.CreateStepper(stepper));
+                stateSpace.Initialize(2, 1, new Matrix(2, 1), stepper: StateSpace.CreateStepper(solver));
                 stateSpace.DerivativeFcn = (states, inputs) =>
                 {
                     // states[0] is position, states[1] is speed
