+++ UnityProject/Assets/Scripts/Core/Interfaces/Mechanical.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 01b548b82b2b6684d9531b74f21658df
-AssemblyDefinitionImporter:
+guid: 58885c3bf1e9280448f87b11c70f1ce6
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
