+++ UnityProject/Assets/Scripts/Core/Solvers/ODESolver.cs
@@ -1,0 +1,54 @@
+using System;
+using SimuNEX.Models;
+using UnityEngine;
+
+namespace SimuNEX.Solvers
+{
+    /// <summary>
+    /// Interface for implementing numerical stepper methods for <see cref="StateSpace"/>.
+    /// </summary>
+    public abstract class ODESolver
+    {
+        /// <summary>
+        /// The step size. Defaults to <see cref="Time.fixedDeltaTime"/>.
+        /// </summary>
+        protected float h = Time.fixedDeltaTime;
+
+        /// <summary>
+        /// The step size for numerical integration.
+        /// </summary>
+        /// <exception cref="ArgumentException"></exception>
+        public float stepSize
+        {
+            get => h;
+
+            set
+            {
+                if (value <= 0)
+                {
+                    throw new ArgumentException("Step size must be positive");
+                }
+
+                if (value < 0.0025f)
+                {
+                    throw new ArgumentException(@"Step size must be at least 0.0025 
+                        to maintain reasonable performance");
+                }
+
+                float remainder = value % 0.005f;
+                if (Math.Abs(remainder) > float.Epsilon)
+                {
+                    throw new ArgumentException("Step size must be a multiple of 0.005");
+                }
+
+                h = value;
+            }
+        }
+
+        /// <summary>
+        /// Numerically integrates the <see cref="StateSpace.DerivativeFunction"/>.
+        /// </summary>
+        /// <param name="ss"><see cref="StateSpace"/> to be integrated.</param>
+        public abstract void Step(StateSpace ss);
+    }
+}
