--- UnityProject/Assets/Scripts/Solvers/SimuNEX.Solvers.asmdef
@@ -1,16 +1,0 @@
-{
-    "name": "SimuNEX.Solvers",
-    "rootNamespace": "",
-    "references": [
-        "GUID:e69e5b0eedcadc743a87e1fc2c3972ed"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
