--- UnityProject/Assets/Scripts/Solvers/SimuNEX.Solvers.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 61af3ef2dd62b4d42a14cd57849a0b54
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
