+++ UnityProject/Assets/Testing/EditMode/Core/ODESolver/SolverTests.Derived.cs
@@ -1,0 +1,16 @@
+using SimuNEX.Solvers;
+
+namespace ODESolverTests
+{
+    public class TestEuler : SolverTests<ForwardEuler>
+    {
+    }
+
+    public class TestHeun : SolverTests<Heun>
+    {
+    }
+
+    public class TestRK4 : SolverTests<RK4>
+    {
+    }
+}
