+++ UnityProject/Assets/Testing/EditMode/Core/StateSpace/FirstOrderTFTests.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
-using static SimuNEX.StateSpaceTypes;
-using static SimuNEX.ODESolvers;
+using SimuNEX.Models;
+using SimuNEX.Solvers;
 
 namespace StateSpaceTests
 {
