--- UnityProject/Assets/Testing/EditMode/Solvers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 69eafca15522373429fc6f8a099481a7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
