--- UnityProject/Assets/Testing/EditMode/Solvers/ODESolver/TestEuler.cs
@@ -1,8 +1,0 @@
-using SimuNEX;
-
-namespace ODESolverTests
-{
-    public class TestEuler : SolverTests<ODESolvers.ForwardEuler>
-    {
-    }
-}
