--- UnityProject/Assets/Testing/EditMode/Solvers/ODESolver/TestHeun.cs
@@ -1,8 +1,0 @@
-using SimuNEX;
-
-namespace ODESolverTests
-{
-    public class TestHeun : SolverTests<ODESolvers.Heun>
-    {
-    }
-}
