--- UnityProject/Assets/Testing/EditMode/Solvers/ODESolver/TestRK4.cs
@@ -1,8 +1,0 @@
-using SimuNEX;
-
-namespace ODESolverTests
-{
-    public class TestRK4 : SolverTests<ODESolvers.RK4>
-    {
-    }
-}
