--- UnityProject/Assets/Testing/EditMode/Solvers/ODESolver/TestRK4.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 740244901b9952648a5390840e787bd7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
